remoteproc: rproc-elf-loader: Add elf load support
Remoteproc framework provides an api to load up the remote processor
image. It is up to individual driver's responsibility to detect the
image type and load accordingly. Current users always assume that
only bin files are loaded directly and copies the image to the
destination. But this is not always true and many remote processor
firmware images are of type ELF. Each section in this ELF can point
to different address within this SoC.

For loading an ELF image, driver's rproc_load api can call the
existing functions in cmd/elf.c. But the problem here is that load
address specified in ELF section is remote processor's view of the
SoC. This address might not accessible or aliased to a different
address for the processor on which U-Boot is running.

So, Introduce support for remoteproc elf loading that gets the cpu
view of the section load addr and load the sections accordingly. In
order to get this cpu view, added a hook inside dm_rpoc_ops called
da_to_va that allows to do the this translation specific to each driver.

This is based on drivers/remoteproc/remoteproc_elf_loader.c in linux kernel

Signed-off-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>