HACK: mmc: am654_sdhci: Wait for 1 second card detect debounce
The 4 bit controllers have an internal timer that takes 1 second to
debounce the card detect signal. It doesn't let software power on the
controller before that time is up.

Therefore, poll on card detect upto a maximum of 2 seconds and only then
continue with powering up the host controller.

TODO: Modify the mmc framework such that this happens at mmc_init time
instead of probe time. To be done in upstream first.

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>