arm: mach-k3: sysfw-loader: Call PM config done callback asap
Call config_pm_done_callback() as soon as all clocks and PM aspects are
setup. This is required because SYSFW may reconfigure clocks after PM
configs are applied leading to boot media being inaccessible as is the
case with HyperBus (bus clock is at wrong frequency), resulting in
failures when loading the RM and Security configs.

Since UART may not be configured at this stage, move printing of SYSFW
version info message to separate function so that it can be invoked by
individual board files after console initialization. Make corresponding
changes to board_init_f() of AM6 and J721E.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>