HACK: arm: dts: k3-j721e: mark rprocs as pd_shared
In j721e devices, rprocs can be booted up from u-boot
and can be attached to in ipc-only mode by the kernel

It is therefore required to request them in shared mode
so that u-boot doesn't call put_device for these devices.

This is a temporary hack and needs to be fixed properly.

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>