HACK: j721e: re-init MMC boot device after sysfw is loaded
One sysfw is loaded, sysfw will configure all PLLs to a known state.
During this time, clocks to the boot media get reset and might observed
a glitch on the boot peripheral. Because of this the boot peripheral
might not work if tries to fetch the next image. To avoid this,
re-initialize the boot peripheral after loading sysfw.

This is the same fix as 49860ebe68 ("HACK: am65x: re-init mmc boot
device after sysfw is loaded").

Acked-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>