mmc: omap_hsmmc: reset FSM for DAT and CMD lines if needed before a new command
commit b2ffa3320ac4b3f040c80768e2b8456bc23aebcc upstream

It sometimes happen that the PSTATE register does not indicate that the
bus is ready when it really is. This usually happens after a mode switch.
In that case it makes sense to reset the FSM handling the CMD and DATA

Also reset the FSMs if the STATE register cannot be cleared. This also
sometimes happens after a mode switch.

Tested-by: Faiz Abbas <faiz_abbas@ti.com>
Reviewed-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>