mmc: add mmc_poll_for_busy() and change the purpose of mmc_send_status()
commit 863d10044aae5ee40a9cf9a76aed62822cf45c30 upstream

mmc_send_status() is currently used to poll the card until it is ready, not
actually returning the status of the card.
Make it return the status and add another function to poll the card.

Also remove the 'extern' declaration in the mmc-private.h header to comply
with the coding standard.

Tested-by: Faiz Abbas <faiz_abbas@ti.com>
Reviewed-by: Faiz Abbas <faiz_abbas@ti.com>
Signed-off-by: Jean-Jacques Hiblot <jjhiblot@ti.com>