dma: ti: k3-udma: Remove coherency check for cache ops
Now that whether or cache ops needs to be done is decided by SoC level
config remove redundant coherency check of cache ops in DMA driver. This
is required for upcomin DMA support for R5 SPL where the core is in the
coherency domain anc cache maintenance ops is required.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>