include: configs: j721e_evm: Add Android boot commands to the environment
Add the following Android boot support commands to the default env:

Android images can be installed and upgraded using a tool called
'fastboot'. To allow U-boot to enter this mode this without user physical
interaction from Android or a previous fastboot session check for a
'dofastboot' variable and enter fastboot if found.

The presence of an Android install on a board can be detected
automatically by checking for a partition labeled 'boot' on the eMMC
device. Add this check command.

Android, like HS Linux, uses FIT images for storing all needed boot
artifacts and their associated configuration/usage information in a
single file, add support for this here.

Acked-by: Praneeth Bajjuri <praneeth@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>