board: ti: am654: Detect PCIEUSB type-A cards
These are identical to SER-PCIEUSBEVM except for
the type of USB connector. They use 3.0 Type-A instead
of 3.0 Micro-AB. They are supposed to work with the
same device tree overlay.

However these cards have 2 spaces at the end of the
card name and are not detected with original code.

Add a new daughtercard entry for this board.

Signed-off-by: Roger Quadros <rogerq@ti.com>