arm: mach-k3: j721e_init: clean up setup_dss_credentials
The code in setup_dss_credentials() is rather ugly, and it has a bug
where it writes to wrong QOS registers, as it uses 'ch' twice as offset,
first adding 4*c, then indexing with ch.

Clean up the whole function and fix the bug.

Also add identical setup for the second master port.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>