remoteproc: ipu: Fixup Kconfig text and add a SoC build dependency
The IPU remoteproc driver is designed to boot the Dual-Cortex M4
IPU remote processors present on the DRA7xx/AM57xx family of SoCs,
and is not applicable for TI K3 SoCs.

Fix up the Kconfig text and add a dependency against DRA7XX Kconfig
symbol. The dependency is limited to DRA7xx/AM57xx SoCS only at the
moment. The support for IPU on OMAP4/OMAP5 SoCs can be added later
on if needed.

Signed-off-by: Suman Anna <s-anna@ti.com>