configs: am65x_evm: Fix Kernel console argument for UBI boot
When performing UBI boot the Kernel commandline 'console' argument was
not configured correctly, leading to an invalid commandline getting
passed to the Kernel. Then, with the Kernel not explicitly knowing its
initial console it would duplicate the first ~50 lines of the boot log
on the serial port.

Fix this issue by using the proper scheme to compile the Kernel arguments
for booting from UBI.

Reviewed-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Andreas Dannenberg <dannenberg@ti.com>