mmc: sdhci: Fix SD card voltage switch
Early in the enumeration process, it is unknown if an attached card is
an SD or an eMMC and the I/O voltage switch to 3.3V never succeeds.

Fix this by always switching voltage if a vqmmc-supply is present but
only do the register configuration if the card IS_SD().

Signed-off-by: Faiz Abbas <faiz_abbas@ti.com>