drivers: net: ti: icssg-prueth: Fix start error path
Currently when start fails due to phy_init we are not
freeing the rx/tx dma channels. This results in errors
in case cable was not connected & later connected and retried.
The error happens only when a ping/dhcp/tftp or any network
related command is tried without the cable physically connected
to the port.

Fix the error path to free up the dma channels.

Signed-off-by: Keerthy <j-keerthy@ti.com>